% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{define_conditions}
\alias{define_conditions}
\title{Combine conditional data-generation rules}
\usage{
define_conditions(...)
}
\arguments{
\item{...}{Objects created by \code{define_condition()}.}
}
\value{
A \code{simulab_condition_spec} base \code{data.frame}.
}
\description{
Combine conditional data-generation rules
}
\examples{
define_conditions(
  define_condition("outcome", condition = "group == 1",
                   formula = "2", variance = "1", distribution = "normal"),
  define_condition("outcome", condition = "group == 0",
                   formula = "0", variance = "1", distribution = "normal")
)
}
