% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalogues.R
\name{global_name_regions}
\alias{global_name_regions}
\title{List global-name regions}
\usage{
global_name_regions()
}
\value{
A character vector of available region identifiers.
}
\description{
List global-name regions
}
\examples{
global_name_regions()
}
