% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning_states.R
\name{learning_state_categories}
\alias{learning_state_categories}
\title{List learning-state categories}
\usage{
learning_state_categories()
}
\value{
A character vector of available category identifiers.
}
\description{
List learning-state categories
}
\examples{
learning_state_categories()
}
