% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas_splines.R
\name{linear_formula}
\alias{linear_formula}
\title{Construct a linear simulation formula}
\usage{
linear_formula(coefficients, variables)
}
\arguments{
\item{coefficients}{Coefficients with an optional intercept.}

\item{variables}{Variable names. When there is one additional coefficient,
the first is the intercept.}
}
\value{
A one-row \code{simulab_formula} base \code{data.frame}.
}
\description{
Construct a linear simulation formula
}
\examples{
linear_formula(coefficients = c(1, 0.5, -0.3), variables = c("x1", "x2"))
}
