% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas_splines.R
\name{mixture_formula}
\alias{mixture_formula}
\title{Construct a mixture simulation formula}
\usage{
mixture_formula(variables, probabilities = NULL)
}
\arguments{
\item{variables}{Variable names or expressions.}

\item{probabilities}{Optional probabilities. Equal probabilities are used by
default.}
}
\value{
A one-row \code{simulab_formula} base \code{data.frame}.
}
\description{
Construct a mixture simulation formula
}
\examples{
mixture_formula(variables = c("a", "b"), probabilities = c(0.3, 0.7))
}
