% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingness.R
\name{define_missingnesses}
\alias{define_missingnesses}
\title{Combine missingness definitions}
\usage{
define_missingnesses(...)
}
\arguments{
\item{...}{Objects created by \code{define_missingness()}.}
}
\value{
A \code{simulab_missing_spec} base \code{data.frame} with one row per target
variable.
}
\description{
Combine missingness definitions
}
\examples{
define_missingnesses(
  define_missingness("outcome", formula = "0.2"),
  define_missingness("baseline", formula = "0.1")
)
}
