% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{define_survival}
\alias{define_survival}
\title{Define a survival process}
\usage{
define_survival(event, formula = 0, scale = 1, shape = 1, transition = 0)
}
\arguments{
\item{event}{Name of the event-time variable.}

\item{formula}{Log-hazard formula.}

\item{scale}{Positive Weibull scale formula.}

\item{shape}{Positive Weibull shape formula in the simstudy parameterization.}

\item{transition}{Time at which this hazard specification begins.}
}
\value{
A one-row \code{simulab_survival_spec} base \code{data.frame}.
}
\description{
Define a survival process
}
\examples{
define_survival("time", formula = -8, shape = 0.3)
}
