% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{define_survivals}
\alias{define_survivals}
\title{Combine survival definitions}
\usage{
define_survivals(...)
}
\arguments{
\item{...}{Objects created by \code{define_survival()}.}
}
\value{
A \code{simulab_survival_spec} base \code{data.frame} with one row per hazard
segment.
}
\description{
Combine survival definitions
}
\examples{
define_survivals(
  define_survival("time_relapse", formula = -8, shape = 0.3),
  define_survival("time_death", formula = -9, shape = 0.3)
)
}
