% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specification.R
\name{define_variables}
\alias{define_variables}
\title{Combine variable definitions}
\usage{
define_variables(...)
}
\arguments{
\item{...}{Objects created by \code{define_variable()} or \code{repeat_variables()}.}
}
\value{
A \code{simulab_spec} base \code{data.frame} with one row per variable.
}
\description{
Combine variable definitions
}
\examples{
define_variables(
  define_variable("age", 40, 100, "normal"),
  define_variable("treated", 0.5, distribution = "binary")
)
}
