% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_structure.R
\name{merge_studies}
\alias{merge_studies}
\title{Merge study data by identifiers}
\usage{
merge_studies(x, y, by, join = c("inner", "full", "left"))
}
\arguments{
\item{x, y}{Base data frames.}

\item{by}{Identifier variables shared by both inputs.}

\item{join}{Join type.}
}
\value{
A \code{simulab_sim} base \code{data.frame} containing the merged data.
}
\description{
Merge study data by identifiers
}
\examples{
x <- data.frame(id = 1:5, baseline = 1:5)
y <- data.frame(id = 1:5, outcome = 6:10)
merge_studies(x, y, by = "id")
}
