% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_structure.R
\name{select_variables}
\alias{select_variables}
\title{Select or remove study variables}
\usage{
select_variables(data, keep = NULL, drop = NULL)
}
\arguments{
\item{data}{Base \code{data.frame}.}

\item{keep}{Variables to retain.}

\item{drop}{Variables to remove.}
}
\value{
A \code{simulab_sim} base \code{data.frame} with the requested variables.
}
\description{
Select or remove study variables
}
\examples{
data <- data.frame(id = 1:5, keep_me = 1:5, drop_me = 6:10)
select_variables(data, keep = c("id", "keep_me"))
select_variables(data, drop = "drop_me")
}
